Worked on rides managements

+++ server/src/main/java/hupiat/scootio/server/accounts/AccountRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AccountRepository extends JpaRepository<AccountEntity, Integer> {
+public interface AccountRepository extends JpaRepository<AccountEntity, Long> {
 
 	Optional<AccountEntity> findByEmail(String email);
